+++ InvaderGame/Shader.cpp
@@ -36,5 +36,5 @@ void Shader::CreateShader(ID3D11Device& device)
 		{"TEXUV", 0, DXGI_FORMAT_R32G32_FLOAT, 0, 12, D3D11_INPUT_PER_VERTEX_DATA, 0}
 	};
 
-	device.CreateInputLayout(&layout[0], layout.size(), compiledVS->GetBufferPointer(), compiledVS->GetBufferSize(), &_inputLayout);
+	device.CreateInputLayout(&layout[0], (UINT)layout.size(), compiledVS->GetBufferPointer(), compiledVS->GetBufferSize(), &_inputLayout);
 }
